refactor xlsx fixture helper facade
